% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calIRGPI.R
\name{calIRGPI}
\alias{calIRGPI}
\title{Calculate immune-related genes pair index (IRGPI) for bladder cancer}
\usage{
calIRGPI(expr = NULL, res.path = getwd())
}
\arguments{
\item{expr}{A numerical expression matrix or data frame with row for gene symbol name and column for sample ID. Note: In principle, the expression profile does not need any normalization, but since the amount of gene expression is affected by the gene length, the original count or the normalized count data may not be suitable for this analysis. It is recommended to provide FPKM or TPM value and log transformation is not necessary.}

\item{res.path}{A string value to indicate the output path for storing the Estimated IRGPI.txt file.}
}
\value{
A data frame stored sample name and estimated IRGPI.
}
\description{
This function calculates an IRGPI risk score individually for bladder cancer only.
}
\examples{
library(BCaller)
load(system.file("extdata", "demo.RData", package = "BCaller", mustWork = TRUE)) # load example data
IRGPI  <- calIRGPI(expr = demo)
head(IRGPI)
}
\references{
Prognosis stratification and personalized treatment in bladder cancer through a robust immune gene-pair based signature
}
\author{
Xiaofan Lu
}
